allow custom hostnames setting for db, elastic, rabbit with envvar
You can not always assume the hostname of the remote services.

For example, in tezos-k8s, I am running everything in a pod, so
I need to set everything to `localhost`.

In order to not disrupt any existing workflow, we set these env
vars to standard values in the Dockerfile; they can be overridden
at runtime. See: https://stackoverflow.com/a/40334819/207209